Feat : Jwt토큰 활용 유저 정보 필요한 기능 구현

+++ src/main/java/com/sparta/springnewsfeed/service/PostService.java
@@ -1,9 +1,12 @@
 package com.sparta.springnewsfeed.service;
 
+import com.sparta.springnewsfeed.exception.InvalidCredentialsException;
 import com.sparta.springnewsfeed.dto.PostRequestDto;
 import com.sparta.springnewsfeed.dto.PostResponseDto;
 import com.sparta.springnewsfeed.entity.Post;
+import com.sparta.springnewsfeed.entity.User;
 import com.sparta.springnewsfeed.repository.PostRepository;
+import com.sparta.springnewsfeed.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -12,9 +15,11 @@ import java.util.List;
 @Service
 public class PostService {
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
-    public PostService(PostRepository postRepository) {
+    public PostService(PostRepository postRepository, UserRepository userRepository) {
         this.postRepository = postRepository;
+        this.userRepository = userRepository;
     }
 
     public PostResponseDto findById(Long postId) {
@@ -24,10 +29,15 @@ public class PostService {
         return new PostResponseDto(post);
     }
 
-    public PostResponseDto createPost(PostRequestDto requestDto) {
+    public PostResponseDto createPost(PostRequestDto requestDto, Long id) {
         Post post = new Post(requestDto);
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new NullPointerException("삭제된 유저 입니다.")
+        );
+        post.setUser(user);
         Post savePost = postRepository.save(post);
-
+        user.getPosts().add(savePost);
+        userRepository.save(user);
         return new PostResponseDto(savePost);
     }
 
@@ -41,15 +51,18 @@ public class PostService {
         return new PostResponseDto(savePost);
     }
 
-    public void deleteById(Long postId) {
+    public void deleteById(Long postId, Long id) {
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new NullPointerException("해당 게시물이 존재하지 않습니다.")
         );
-        postRepository.delete(post);
+        if (post.getUser().getId().equals(id)) {
+            postRepository.delete(post);
+        } else {
+            throw new InvalidCredentialsException("권한이 없습니다.");
+        }
     }
 
-    public List<PostResponseDto> findMyPost() {
-        Long userid = 123L;
+    public List<PostResponseDto> findMyPost(Long userid) {
         List<Post> myPosts = postRepository.findByUserId(userid).orElseThrow(
                 () -> new NullPointerException("작성한 게시물이 없습니다.")
         );
